w3c-reports/12092014/mac/runned-tests/XMLHttpRequest/interfaces-expected.txt

CONSOLE MESSAGE: line 188: callback not yet supported
XMLHttpRequest IDL tests


FAIL XMLHttpRequestUpload interface: existence and properties of interface object assert_equals: prototype of window's property "XMLHttpRequestUpload" is not Function.prototype expected (function) function "function () {
    [native code]
}" but got (object) object "[object Object]"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/resources/idlharness.js:991:22
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:947:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequestUpload interface object length 
FAIL XMLHttpRequestUpload interface: existence and properties of interface prototype object assert_false: XMLHttpRequestUpload.prototype is enumerable expected false got true(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_false@http://localhost:8800/resources/testharness.js:306:15
http://localhost:8800/resources/idlharness.js:1094:21
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1071:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestUpload interface: existence and properties of interface prototype object's "constructor" property assert_true: XMLHttpRequestUpload.prototype.constructor is not writable expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1183:20
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1159:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequestUpload must be primary interface of (new XMLHttpRequest()).upload 
PASS Stringification of (new XMLHttpRequest()).upload 
FAIL XMLHttpRequestEventTarget interface: (new XMLHttpRequest()).upload must inherit property "onloadstart" with the proper type (0) assert_inherits: property "onloadstart" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestEventTarget interface: (new XMLHttpRequest()).upload must inherit property "onprogress" with the proper type (1) assert_inherits: property "onprogress" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestEventTarget interface: (new XMLHttpRequest()).upload must inherit property "onabort" with the proper type (2) assert_inherits: property "onabort" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestEventTarget interface: (new XMLHttpRequest()).upload must inherit property "onerror" with the proper type (3) assert_inherits: property "onerror" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestEventTarget interface: (new XMLHttpRequest()).upload must inherit property "onload" with the proper type (4) assert_inherits: property "onload" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestEventTarget interface: (new XMLHttpRequest()).upload must inherit property "ontimeout" with the proper type (5) assert_inherits: property "ontimeout" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestEventTarget interface: (new XMLHttpRequest()).upload must inherit property "onloadend" with the proper type (6) assert_inherits: property "onloadend" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS EventTarget interface: (new XMLHttpRequest()).upload must inherit property "addEventListener" with the proper type (0) 
FAIL EventTarget interface: calling addEventListener(DOMString,EventListener,boolean) on (new XMLHttpRequest()).upload with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1640:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS EventTarget interface: (new XMLHttpRequest()).upload must inherit property "removeEventListener" with the proper type (1) 
FAIL EventTarget interface: calling removeEventListener(DOMString,EventListener,boolean) on (new XMLHttpRequest()).upload with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1640:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS EventTarget interface: (new XMLHttpRequest()).upload must inherit property "dispatchEvent" with the proper type (2) 
PASS EventTarget interface: calling dispatchEvent(Event) on (new XMLHttpRequest()).upload with too few arguments must throw TypeError 
FAIL XMLHttpRequest interface: existence and properties of interface object assert_equals: prototype of window's property "XMLHttpRequest" is not Function.prototype expected (function) function "function () {
    [native code]
}" but got (object) object "[object Object]"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/resources/idlharness.js:991:22
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:947:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequest interface object length 
FAIL XMLHttpRequest interface: existence and properties of interface prototype object assert_false: XMLHttpRequest.prototype is enumerable expected false got true(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_false@http://localhost:8800/resources/testharness.js:306:15
http://localhost:8800/resources/idlharness.js:1094:21
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1071:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: existence and properties of interface prototype object's "constructor" property assert_true: XMLHttpRequest.prototype.constructor is not writable expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1183:20
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1159:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: attribute onreadystatechange assert_true: The prototype object must have a property "onreadystatechange" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequest interface: constant UNSENT on interface object 
PASS XMLHttpRequest interface: constant UNSENT on interface prototype object 
PASS XMLHttpRequest interface: constant OPENED on interface object 
PASS XMLHttpRequest interface: constant OPENED on interface prototype object 
PASS XMLHttpRequest interface: constant HEADERS_RECEIVED on interface object 
PASS XMLHttpRequest interface: constant HEADERS_RECEIVED on interface prototype object 
PASS XMLHttpRequest interface: constant LOADING on interface object 
PASS XMLHttpRequest interface: constant LOADING on interface prototype object 
PASS XMLHttpRequest interface: constant DONE on interface object 
PASS XMLHttpRequest interface: constant DONE on interface prototype object 
FAIL XMLHttpRequest interface: attribute readyState assert_true: The prototype object must have a property "readyState" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequest interface: operation open(ByteString,DOMString) 
FAIL XMLHttpRequest interface: operation open(ByteString,DOMString,boolean,DOMString,DOMString) assert_equals: property has wrong .length expected 3 but got 2(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1336:22
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequest interface: operation setRequestHeader(ByteString,ByteString) 
FAIL XMLHttpRequest interface: attribute timeout assert_true: The prototype object must have a property "timeout" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: attribute withCredentials assert_true: The prototype object must have a property "withCredentials" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: attribute upload assert_true: The prototype object must have a property "upload" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequest interface: operation send([object Object],[object Object],[object Object],[object Object],[object Object],[object Object]) 
PASS XMLHttpRequest interface: operation abort() 
FAIL XMLHttpRequest interface: attribute status assert_true: The prototype object must have a property "status" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: attribute statusText assert_true: The prototype object must have a property "statusText" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequest interface: operation getResponseHeader(ByteString) 
PASS XMLHttpRequest interface: operation getAllResponseHeaders() 
PASS XMLHttpRequest interface: operation overrideMimeType(DOMString) 
FAIL XMLHttpRequest interface: attribute responseType assert_true: The prototype object must have a property "responseType" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: attribute response assert_true: The prototype object must have a property "response" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: attribute responseText assert_true: The prototype object must have a property "responseText" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: attribute responseXML assert_true: The prototype object must have a property "responseXML" expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1265:24
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_attribute@http://localhost:8800/resources/idlharness.js:1251:9
test_members@http://localhost:8800/resources/idlharness.js:1450:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequest must be primary interface of new XMLHttpRequest() 
PASS Stringification of new XMLHttpRequest() 
FAIL XMLHttpRequest interface: new XMLHttpRequest() must inherit property "onreadystatechange" with the proper type (0) assert_inherits: property "onreadystatechange" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequest interface: new XMLHttpRequest() must inherit property "UNSENT" with the proper type (1) 
PASS XMLHttpRequest interface: new XMLHttpRequest() must inherit property "OPENED" with the proper type (2) 
PASS XMLHttpRequest interface: new XMLHttpRequest() must inherit property "HEADERS_RECEIVED" with the proper type (3) 
PASS XMLHttpRequest interface: new XMLHttpRequest() must inherit property "LOADING" with the proper type (4) 
PASS XMLHttpRequest interface: new XMLHttpRequest() must inherit property "DONE" with the proper type (5) 
FAIL XMLHttpRequest interface: new XMLHttpRequest() must inherit property "readyState" with the proper type (6) assert_inherits: property "readyState" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequest interface: new XMLHttpRequest() must inherit property "open" with the proper type (7) 
PASS XMLHttpRequest interface: calling open(ByteString,DOMString) on new XMLHttpRequest() with too few arguments must throw TypeError 
PASS XMLHttpRequest interface: new XMLHttpRequest() must inherit property "open" with the proper type (8) 
FAIL XMLHttpRequest interface: calling open(ByteString,DOMString,boolean,DOMString,DOMString) on new XMLHttpRequest() with too few arguments must throw TypeError assert_throws: Called with 2 arguments function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1640:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequest interface: new XMLHttpRequest() must inherit property "setRequestHeader" with the proper type (9) 
PASS XMLHttpRequest interface: calling setRequestHeader(ByteString,ByteString) on new XMLHttpRequest() with too few arguments must throw TypeError 
FAIL XMLHttpRequest interface: new XMLHttpRequest() must inherit property "timeout" with the proper type (10) assert_inherits: property "timeout" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: new XMLHttpRequest() must inherit property "withCredentials" with the proper type (11) assert_inherits: property "withCredentials" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: new XMLHttpRequest() must inherit property "upload" with the proper type (12) assert_inherits: property "upload" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequest interface: new XMLHttpRequest() must inherit property "send" with the proper type (13) 
PASS XMLHttpRequest interface: calling send([object Object],[object Object],[object Object],[object Object],[object Object],[object Object]) on new XMLHttpRequest() with too few arguments must throw TypeError 
PASS XMLHttpRequest interface: new XMLHttpRequest() must inherit property "abort" with the proper type (14) 
FAIL XMLHttpRequest interface: new XMLHttpRequest() must inherit property "status" with the proper type (15) assert_inherits: property "status" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: new XMLHttpRequest() must inherit property "statusText" with the proper type (16) assert_inherits: property "statusText" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS XMLHttpRequest interface: new XMLHttpRequest() must inherit property "getResponseHeader" with the proper type (17) 
PASS XMLHttpRequest interface: calling getResponseHeader(ByteString) on new XMLHttpRequest() with too few arguments must throw TypeError 
PASS XMLHttpRequest interface: new XMLHttpRequest() must inherit property "getAllResponseHeaders" with the proper type (18) 
PASS XMLHttpRequest interface: new XMLHttpRequest() must inherit property "overrideMimeType" with the proper type (19) 
PASS XMLHttpRequest interface: calling overrideMimeType(DOMString) on new XMLHttpRequest() with too few arguments must throw TypeError 
FAIL XMLHttpRequest interface: new XMLHttpRequest() must inherit property "responseType" with the proper type (20) assert_inherits: property "responseType" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: new XMLHttpRequest() must inherit property "response" with the proper type (21) assert_inherits: property "response" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: new XMLHttpRequest() must inherit property "responseText" with the proper type (22) assert_inherits: property "responseText" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequest interface: new XMLHttpRequest() must inherit property "responseXML" with the proper type (23) assert_inherits: property "responseXML" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestEventTarget interface: new XMLHttpRequest() must inherit property "onloadstart" with the proper type (0) assert_inherits: property "onloadstart" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestEventTarget interface: new XMLHttpRequest() must inherit property "onprogress" with the proper type (1) assert_inherits: property "onprogress" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestEventTarget interface: new XMLHttpRequest() must inherit property "onabort" with the proper type (2) assert_inherits: property "onabort" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestEventTarget interface: new XMLHttpRequest() must inherit property "onerror" with the proper type (3) assert_inherits: property "onerror" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestEventTarget interface: new XMLHttpRequest() must inherit property "onload" with the proper type (4) assert_inherits: property "onload" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestEventTarget interface: new XMLHttpRequest() must inherit property "ontimeout" with the proper type (5) assert_inherits: property "ontimeout" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL XMLHttpRequestEventTarget interface: new XMLHttpRequest() must inherit property "onloadend" with the proper type (6) assert_inherits: property "onloadend" found on object expected in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:548:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS EventTarget interface: new XMLHttpRequest() must inherit property "addEventListener" with the proper type (0) 
FAIL EventTarget interface: calling addEventListener(DOMString,EventListener,boolean) on new XMLHttpRequest() with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1640:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS EventTarget interface: new XMLHttpRequest() must inherit property "removeEventListener" with the proper type (1) 
FAIL EventTarget interface: calling removeEventListener(DOMString,EventListener,boolean) on new XMLHttpRequest() with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1640:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS EventTarget interface: new XMLHttpRequest() must inherit property "dispatchEvent" with the proper type (2) 
PASS EventTarget interface: calling dispatchEvent(Event) on new XMLHttpRequest() with too few arguments must throw TypeError 
FAIL FormData interface: existence and properties of interface object assert_equals: prototype of window's property "FormData" is not Function.prototype expected (function) function "function () {
    [native code]
}" but got (object) object "[object Object]"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/resources/idlharness.js:991:22
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:947:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS FormData interface object length 
FAIL FormData interface: existence and properties of interface prototype object assert_false: FormData.prototype is enumerable expected false got true(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_false@http://localhost:8800/resources/testharness.js:306:15
http://localhost:8800/resources/idlharness.js:1094:21
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1071:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: existence and properties of interface prototype object's "constructor" property assert_true: FormData.prototype.constructor is not writable expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/resources/idlharness.js:1183:20
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_self@http://localhost:8800/resources/idlharness.js:1159:9
test@http://localhost:8800/resources/idlharness.js:582:23
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: operation append(DOMString,Blob,DOMString) assert_equals: property has wrong .length expected 2 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1336:22
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: operation append(DOMString,DOMString) assert_equals: property has wrong .length expected 2 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/resources/idlharness.js:1336:22
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: operation delete(DOMString) assert_own_property: interface prototype object missing non-static operation expected property "delete" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1312:32
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: operation get(DOMString) assert_own_property: interface prototype object missing non-static operation expected property "get" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1312:32
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: operation getAll(DOMString) assert_own_property: interface prototype object missing non-static operation expected property "getAll" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1312:32
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: operation has(DOMString) assert_own_property: interface prototype object missing non-static operation expected property "has" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1312:32
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: operation set(DOMString,Blob,DOMString) assert_own_property: interface prototype object missing non-static operation expected property "set" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1312:32
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: operation set(DOMString,DOMString) assert_own_property: interface prototype object missing non-static operation expected property "set" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:521:19
http://localhost:8800/resources/idlharness.js:1312:32
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_member_operation@http://localhost:8800/resources/idlharness.js:1282:9
test_members@http://localhost:8800/resources/idlharness.js:1458:43
test@http://localhost:8800/resources/idlharness.js:591:22
test@http://localhost:8800/resources/idlharness.js:311:32
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS FormData must be primary interface of new FormData() 
PASS Stringification of new FormData() 
PASS FormData interface: new FormData() must inherit property "append" with the proper type (0) 
FAIL FormData interface: calling append(DOMString,Blob,DOMString) on new FormData() with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1640:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS FormData interface: new FormData() must inherit property "append" with the proper type (1) 
FAIL FormData interface: calling append(DOMString,DOMString) on new FormData() with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1640:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: new FormData() must inherit property "delete" with the proper type (2) assert_inherits: property "delete" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: calling delete(DOMString) on new FormData() with too few arguments must throw TypeError assert_inherits: property "delete" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1627:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: new FormData() must inherit property "get" with the proper type (3) assert_inherits: property "get" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: calling get(DOMString) on new FormData() with too few arguments must throw TypeError assert_inherits: property "get" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1627:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: new FormData() must inherit property "getAll" with the proper type (4) assert_inherits: property "getAll" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: calling getAll(DOMString) on new FormData() with too few arguments must throw TypeError assert_inherits: property "getAll" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1627:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: new FormData() must inherit property "has" with the proper type (5) assert_inherits: property "has" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: calling has(DOMString) on new FormData() with too few arguments must throw TypeError assert_inherits: property "has" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1627:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: new FormData() must inherit property "set" with the proper type (6) assert_inherits: property "set" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: calling set(DOMString,Blob,DOMString) on new FormData() with too few arguments must throw TypeError assert_inherits: property "set" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1627:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: new FormData() must inherit property "set" with the proper type (7) assert_inherits: property "set" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: calling set(DOMString,DOMString) on new FormData() with too few arguments must throw TypeError assert_inherits: property "set" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1627:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS FormData must be primary interface of new FormData(form) 
PASS Stringification of new FormData(form) 
PASS FormData interface: new FormData(form) must inherit property "append" with the proper type (0) 
FAIL FormData interface: calling append(DOMString,Blob,DOMString) on new FormData(form) with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1640:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
PASS FormData interface: new FormData(form) must inherit property "append" with the proper type (1) 
FAIL FormData interface: calling append(DOMString,DOMString) on new FormData(form) with too few arguments must throw TypeError assert_throws: Called with 0 arguments function "function () {
    [native code]
}" did not throw(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_throws@http://localhost:8800/resources/testharness.js:583:19
http://localhost:8800/resources/idlharness.js:1640:34
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: new FormData(form) must inherit property "delete" with the proper type (2) assert_inherits: property "delete" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: calling delete(DOMString) on new FormData(form) with too few arguments must throw TypeError assert_inherits: property "delete" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1627:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: new FormData(form) must inherit property "get" with the proper type (3) assert_inherits: property "get" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: calling get(DOMString) on new FormData(form) with too few arguments must throw TypeError assert_inherits: property "get" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1627:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: new FormData(form) must inherit property "getAll" with the proper type (4) assert_inherits: property "getAll" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: calling getAll(DOMString) on new FormData(form) with too few arguments must throw TypeError assert_inherits: property "getAll" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1627:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: new FormData(form) must inherit property "has" with the proper type (5) assert_inherits: property "has" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: calling has(DOMString) on new FormData(form) with too few arguments must throw TypeError assert_inherits: property "has" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1627:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: new FormData(form) must inherit property "set" with the proper type (6) assert_inherits: property "set" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: calling set(DOMString,Blob,DOMString) on new FormData(form) with too few arguments must throw TypeError assert_inherits: property "set" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1627:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: new FormData(form) must inherit property "set" with the proper type (7) assert_inherits: property "set" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1586:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1581:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
FAIL FormData interface: calling set(DOMString,DOMString) on new FormData(form) with too few arguments must throw TypeError assert_inherits: property "set" not found in prototype chain(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
http://localhost:8800/resources/testharness.js:553:19
http://localhost:8800/resources/idlharness.js:1627:36
[native code]
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
test_interface_of@http://localhost:8800/resources/idlharness.js:1622:17
test_object@http://localhost:8800/resources/idlharness.js:1507:44
http://localhost:8800/resources/idlharness.js:316:47
[native code]
forEach@[native code]
test@http://localhost:8800/resources/idlharness.js:314:39
global code@http://localhost:8800/XMLHttpRequest/interfaces.html:115:14)
